[Portfolio] Refactor creation
