+++ octobot_trading/portfolios/portfolio_factory.pxd
@@ -1,0 +1,20 @@
+# cython: language_level=3
+#  Drakkar-Software OctoBot-Trading
+#  Copyright (c) Drakkar-Software, All rights reserved.
+#
+#  This library is free software; you can redistribute it and/or
+#  modify it under the terms of the GNU Lesser General Public
+#  License as published by the Free Software Foundation; either
+#  version 3.0 of the License, or (at your option) any later version.
+#
+#  This library is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+#  Lesser General License for more details.
+#
+#  You should have received a copy of the GNU Lesser General Public
+#  License along with this library.
+from octobot_trading.data.portfolio cimport Portfolio
+from octobot_trading.exchanges.exchange_manager cimport ExchangeManager
+
+cpdef Portfolio create_portfolio_from_exchange_manager(ExchangeManager exchange_manager)
