+++ octobot_trading/portfolios/sub_portfolio.py
