+++ setup.py
@@ -66,13 +66,11 @@ packages_list = ["octobot_trading.util.initializable",
                  "octobot_trading.producers.simulator.prices_updater_simulator",
                  "octobot_trading.producers.simulator.recent_trade_updater_simulator",
                  "octobot_trading.producers.simulator.ticker_updater_simulator",
-                 "octobot_trading.data.margin_portfolio",
                  "octobot_trading.data.order",
                  "octobot_trading.data.position",
                  "octobot_trading.data.trade",
                  "octobot_trading.data.portfolio",
                  "octobot_trading.data.portfolio_profitability",
-                 "octobot_trading.data.sub_portfolio",
                  "octobot_trading.data_adapters.candles_adapter",
                  "octobot_trading.data_manager.candles_manager",
                  "octobot_trading.data_manager.funding_manager",
@@ -108,6 +106,11 @@ packages_list = ["octobot_trading.util.initializable",
                  "octobot_trading.orders.types.trailing.trailing_stop_order",
                  "octobot_trading.orders.types.trailing.trailing_stop_limit_order",
                  "octobot_trading.orders.types.unknown_order",
+                 "octobot_trading.portfolios.portfolio_factory",
+                 "octobot_trading.portfolios.sub_portfolio",
+                 "octobot_trading.portfolios.types.future_portfolio",
+                 "octobot_trading.portfolios.types.margin_portfolio",
+                 "octobot_trading.portfolios.types.spot_portfolio",
                  "octobot_trading.traders.trader",
                  "octobot_trading.traders.trader_simulator",
                  "octobot_trading.trades.trade_factory",
